+++ app/src/main/java/ru/yandex/yamblz/ui/recycler/decorators/ColorsItemDecoration.java
@@ -1,4 +1,4 @@
-package ru.yandex.yamblz.ui.adapters;
+package ru.yandex.yamblz.ui.recycler.decorators;
 
 import android.graphics.Canvas;
 import android.graphics.Paint;
@@ -6,6 +6,8 @@ import android.graphics.Path;
 import android.graphics.Rect;
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import ru.yandex.yamblz.ui.recycler.adapters.ColorsAdapter;
 
 public class ColorsItemDecoration extends RecyclerView.ItemDecoration {
